requirements.txt
google-generativeai
requests
beautifulsoup4
streamlit
praw